+++ test/mitmproxy/proxy/test_mode_servers.py
@@ -21,12 +21,41 @@ def test_make():
         assert inst.log_desc
 
 
+async def test_last_exception_and_running(monkeypatch):
+    manager = MagicMock()
+    err = ValueError("something else")
+
+    async def _raise(*_):
+        nonlocal err
+        raise err
+
+    with taddons.context():
+
+        inst1 = ServerInstance.make("regular@127.0.0.1:0", manager)
+        await inst1.start()
+        assert inst1.last_exception is None
+        assert inst1.is_running
+        monkeypatch.setattr(inst1._server, "wait_closed", _raise)
+        with pytest.raises(type(err), match=str(err)):
+            await inst1.stop()
+        assert inst1.last_exception is err
+
+        monkeypatch.setattr(asyncio, "start_server", _raise)
+        inst2 = ServerInstance.make("regular@127.0.0.1:0", manager)
+        assert inst2.last_exception is None
+        with pytest.raises(type(err), match=str(err)):
+            await inst2.start()
+        assert inst2.last_exception is err
+        assert not inst1.is_running
+
+
 async def test_tcp_start_stop():
     manager = MagicMock()
 
     with taddons.context() as tctx:
         inst = ServerInstance.make("regular@127.0.0.1:0", manager)
         await inst.start()
+        assert inst.last_exception is None
         assert await tctx.master.await_log("proxy listening")
 
         host, port, *_ = inst.listen_addrs[0]
@@ -38,20 +67,19 @@ async def test_tcp_start_stop():
         assert await tctx.master.await_log("client disconnect")
 
         await inst.stop()
-        assert await tctx.master.await_log("Stopped regular proxy server.")
+        assert await tctx.master.await_log("stopped HTTP(S) proxy")
 
 
 async def test_tcp_start_error():
     manager = MagicMock()
 
+    server = await asyncio.start_server(MagicMock(), host="127.0.0.1", port=0, reuse_address=False)
+    port = server.sockets[0].getsockname()[1]
+
     with taddons.context() as tctx:
-        inst = ServerInstance.make("regular@127.0.0.1:0", manager)
-        await inst.start()
-        assert await tctx.master.await_log("proxy listening")
-        port = inst.listen_addrs[0][1]
-        inst2 = ServerInstance.make(f"regular@127.0.0.1:{port}", manager)
+        inst = ServerInstance.make(f"regular@127.0.0.1:{port}", manager)
         with pytest.raises(OSError, match=f"proxy failed to listen on 127\\.0\\.0\\.1:{port}"):
-            await inst2.start()
+            await inst.start()
         tctx.options.listen_host = "127.0.0.1"
         tctx.options.listen_port = port
         inst3 = ServerInstance.make(f"regular", manager)